+++ lib/src/models/scan_filter.dart
@@ -1,4 +1,4 @@
-import 'dart:typed_data';
+import 'package:universal_ble/src/models/manufacturer_data_filter.dart';
 
 class ScanFilter {
   List<String> withServices;
@@ -14,28 +14,5 @@ class ScanFilter {
   @override
   String toString() {
     return 'ScanFilter(withServices: $withServices, withManufacturerData: $withManufacturerData, withNamePrefix: $withNamePrefix)';
-  }
-}
-
-class ManufacturerDataFilter {
-  int companyIdentifier;
-
-  // Mask and data must be of same length
-  Uint8List? payload;
-
-  /// For any bit in the mask, set it the 1 if it needs to match
-  /// the one in manufacturer data, otherwise set it to 0.
-  /// The 'mask' must have the same length as 'data'.
-  Uint8List? mask;
-
-  ManufacturerDataFilter({
-    required this.companyIdentifier,
-    this.payload,
-    this.mask,
-  });
-
-  @override
-  String toString() {
-    return 'ManufacturerDataFilter(companyIdentifier: $companyIdentifier, payload: $payload, mask: $mask)';
   }
 }